% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_NEE.R
\name{analyze_NEE}
\alias{analyze_NEE}
\title{Analyze under NEE assumptions}
\usage{
analyze_NEE(
  data,
  brange = c(0, 0),
  design = "full",
  weights = NULL,
  contrast = "logRR"
)
}
\arguments{
\item{data}{data frame}

\item{brange}{numeric vector}

\item{design}{string}

\item{weights}{numeric vector}

\item{contrast}{string}
}
\value{
list
}
\description{
Analyze data assuming no early treatment effects
}
